requirements.txt
Django
django-filter
djangorestframework
freeze
Markdown
six
twilio
